refactor: merges content of Tv2CasparCgGraphicsTimelineObjectFactory and  Tv2CasparCgTimelineObjectFactory
