+++ src/blueprints/tv2/timeline-object-factories/interfaces/tv2-graphics-timeline-object-factory.ts
@@ -1,6 +1,9 @@
 import { Tv2BlueprintConfiguration } from '../../value-objects/tv2-blueprint-configuration'
 import { TimelineObject } from '../../../../model/entities/timeline-object'
-import { Tv2GraphicsData } from '../../value-objects/tv2-action-manifest-data'
+import {
+  Tv2FullscreenGraphicsManifestData,
+  Tv2OverlayGraphicsManifestData
+} from '../../value-objects/tv2-action-manifest-data'
 
 export interface Tv2GraphicsTimelineObjectFactory {
   createThemeOutTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, duration: number): TimelineObject
@@ -8,7 +11,7 @@ export interface Tv2GraphicsTimelineObjectFactory {
   createContinueGraphicsTimelineObject(duration: number): TimelineObject
   createClearGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, duration: number): TimelineObject
   createAllOutGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, duration: number): TimelineObject
-  createFullGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, graphicsData: Tv2GraphicsData): TimelineObject
-  createIdentGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, graphicsData: Tv2GraphicsData): TimelineObject
-  createLowerThirdGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, graphicsData: Tv2GraphicsData): TimelineObject
+  createFullGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, graphicsData: Tv2FullscreenGraphicsManifestData): TimelineObject
+  createIdentGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, graphicsData: Tv2OverlayGraphicsManifestData): TimelineObject
+  createLowerThirdGraphicsTimelineObject(blueprintConfiguration: Tv2BlueprintConfiguration, graphicsData: Tv2OverlayGraphicsManifestData): TimelineObject
 }
